pyproject.toml
[project]
name = "async-titiler"
description = "TiTiler but trully ==async."
readme = "README.md"
requires-python = ">=3.11"
license = {file = "LICENSE"}
authors = [
    {name = "Vincent Sarago", email = "vincent@developmentseed.com"},
]
keywords = [
    "TiTiler",
    "async",
]
classifiers = [
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Scientific/Engineering :: GIS",
]
dynamic = ["version"]
dependencies = [
    "orjson",
    "async-cache",
    "titiler-core>=2.0,<2.1",
    "async-geotiff>=0.3,<0.5",
    "zarr~=3.0",
    "obstore",
    "rio-tiler@git+https://github.com/cogeotiff/rio-tiler.git@feat/geozarr-reader",
    "starlette-cramjam>=0.4,<0.6",
    "pydantic-settings~=2.0",
]
[project.optional-dependencies]
server = [
    "uvicorn[standard]>=0.12.0",
]

[dependency-groups]
dev = [
    "pre-commit",
    "pytest",
    "pytest-cov",
    "pytest-asyncio",
    "httpx",
]

[tool.hatch.version]
path = "async_titiler/__init__.py"

[tool.hatch.build.targets.sdist]
only-include = ["async_titiler"]

[tool.hatch.build.targets.wheel]
only-include = ["async_titiler"]

[tool.hatch.metadata]
allow-direct-references = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.coverage.run]
branch = true
parallel = true

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.isort]
profile = "black"
known_first_party = ["titiler"]
known_third_party = [
    "rasterio",
    "morecantile",
    "rio_tiler",
    "cogeo_mosaic",
    "geojson_pydantic",
]
default_section = "THIRDPARTY"

[tool.ruff.lint]
select = [
    "D1",  # pydocstyle errors
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # flake8
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "B905",  # ignore zip() without an explicit strict= parameter, only support with python >3.10
]

[tool.mypy]
no_implicit_optional = true
strict_optional = true
namespace_packages = true
explicit_package_bases = true

[tool.bumpversion]
current_version = "1.1.2"
search = "{current_version}"
replace = "{new_version}"
regex = false
tag = false
commit = true
allow_dirty = false
tag_name = "{new_version}"

[[tool.bumpversion.files]]
filename = "async_titiler/__init__.py"
search = '__version__ = "{current_version}"'
replace = '__version__ = "{new_version}"'